Move entry checking to disk_handler
