+++ Main.py
@@ -60,6 +60,8 @@ class MangaCabinet(QWidget):
         self.tag_view = None
         self.styles = load_styles(self.style_path)
         self.settings = Options.load_settings(self.settings_file)
+        self.disk_handler = DiskHandler(self)
+        self.browser_handler = BrowserHandler(self)
 
         self.entry_to_index = {}
         self.tag_data = TagData()
@@ -68,8 +70,6 @@ class MangaCabinet(QWidget):
 
         self.thumbnail_manager = ThumbnailManager(self, self.data, self.download_thumbnails, self.tags_to_blur)
         self.thumbnail_manager.startEnsuring.emit()
-        self.browser_handler = BrowserHandler(self)
-        self.disk_handler = DiskHandler(self)
         self.init_ui()
         self.show()
         self.check_for_backups_and_recovery_files()
@@ -192,7 +192,7 @@ class MangaCabinet(QWidget):
             self.all_artists.update(entry.artist)
             self.tag_data.update_with_entry(entry)
 
-        self.check_entries_disk_locations(self.data)
+        self.disk_handler.check_entries_disk_locations(self.data)
 
     def addNewData(self, newData: list[MangaEntry]):
         self.data = newData + self.data
@@ -202,59 +202,12 @@ class MangaCabinet(QWidget):
         for entry in newData:
             self.all_artists.update(entry.artist)
             self.tag_data.update_with_entry(entry)
-            self.check_entry_disk_location(entry)
+            self.disk_handler.check_entry_disk_location(entry)
 
         self.is_data_modified = True
         self.logger.info(f"Updated data with the following entries: {', '.join([entry.id for entry in newData])}")
         self.dataUpdated.emit(newData)
         self.search_bar_handler.update_list()
-
-    def check_entry_disk_location(self, entry, loose_check=False):
-        # If entry doesn't have a defined filesystem location and default is defined, try to find it
-        if self.settings[default_manga_loc] and not entry.disk_location(loose_check=loose_check):
-            # Special handling for path here because I want to prevent backslash usage in data json
-            entry_path = Path(self.settings[default_manga_loc]) / entry.id
-            if entry_path.exists():
-                # TODO: Streamline save system
-                entry.filesystem_location = str(entry_path).replace('\\', '/')
-                self.is_data_modified = True
-                self.logger.debug(f"{entry.id}: filesystem_location was updated with: {entry.filesystem_location}")
-
-    def check_entries_disk_locations(self, entries, loose_check=True):
-        """
-        Efficiently checks and updates filesystem locations for multiple entries
-        by minimizing disk I/O with a single directory read.
-        """
-        if not self.settings[default_manga_loc]:
-            return
-        manga_loc_path = Path(self.settings[default_manga_loc])
-
-        existing_paths = os.listdir(manga_loc_path)
-
-        # Iterate through each entry in the provided list
-        for entry in entries:
-            cur_loc = entry.filesystem_location
-            cur_loc_path = Path(cur_loc) if cur_loc else None
-
-            if loose_check:
-                # Catches location not written or written but id changed mostly
-                check_condition = (not cur_loc_path or
-                                   (cur_loc_path.is_relative_to(manga_loc_path) and
-                                    cur_loc_path.name not in existing_paths))
-            else:
-                # Catches not written or not existing (does an IO operation, costly, only done rarely)
-                check_condition = (not cur_loc_path or
-                                   not cur_loc_path.exists())
-
-            if check_condition:
-                # Construct the path for the entry
-                entry_path = manga_loc_path / entry.id
-
-                if entry_path.name in existing_paths:
-                    # TODO: Streamline save system
-                    entry.filesystem_location = str(entry_path).replace('\\', '/')
-                    self.is_data_modified = True
-                    self.logger.debug(f"{entry.id}: filesystem_location was updated with: {entry.filesystem_location}")
 
     def open_tab_from_entry(self, entry: MangaEntry):
         if not self.browser_handler.unsupported:
